+++ .github/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <h3>New features since last release</h3>
 
-* Added the ``CnotRingLayers`` template, which is a simple layer architecture 
+* Added the ``BasicEntanglerLayers`` template, which is a simple layer architecture 
   of rotations and CNOT nearest-neighbour entanglers. 
   [(#555)](https://github.com/XanaduAI/pennylane/pull/555)